finished 3

+++ Binomial.java
@@ -24,14 +24,14 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		long[][] memo = new long[n + 1][k + 1];
 		for (int i = 0; i <= n; i++) {
 			for (int j = 0; j <= k; j++) {
 				memo[i][j] = -1;
 			}
 		}
-		return (int) binomial(n, k, memo);
+		return (long) binomial(n, k, memo);
 	}
 
 	private static long binomial(int n, int k, long[][] memo) {
